fix: scanner module > confirm before delete page
git-svn-id: https://win-lhoa5a3q3i7/svn/front-end-backup/trunk@4871 3f25d8b1-e65f-8d48-bd54-75b34d8be7f1

+++ src/modules/scanner-module/controllers/scannerCtrl.js
@@ -107,9 +107,13 @@ module.exports = function (app) {
             return self.cc.getDocument().pages.length > 1;
         };
         self.deletePage = function (page) {
-            self.cc.getDocument().deletePage(page, function (pageAfter) {
-                self.showPage(pageAfter);
-            });
+            debugger;
+            dialog.confirmMessage(langService.get('confirm_delete_msg'))
+                .then(function(){
+                    self.cc.getDocument().deletePage(page, function (pageAfter) {
+                        self.showPage(pageAfter);
+                    });
+                });
         };
         self.enableProperty = function (property) {
             self.disableProperties[property] = false;
